api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pg8000
paho-mqtt
scikit-learn
joblib
python-dotenv
httpx
pandas
